+++ scripts/esbuild.config.ts
@@ -1,31 +1,76 @@
+import postCssPlugin from 'esbuild-style-plugin'
 import type { BuildOptions } from "esbuild";
+
+const banner = `
+/*
+MIT License
+
+Copyright (c) 2025 MolikoDev
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+*/
+`
 
 const config: BuildOptions = {
   entryPoints: [
     "./src/extension.ts",
-    "./src/webview/App.tsx"
+    "./src/webview/App.tsx",
+    "./src/tailwind.css",
+    "./src/common/highlight.min.cjs"
   ],
+  //minify: true,
   bundle: true,
-  platform: "node",
-  target: "node12",
+  platform: "node", // "node" | "browser"
+  target: "node18",
   outdir: "./dist",
   outbase: "./src",
   outExtension: {
     ".js": ".cjs",
+    ".css": ".css"
+  },
+  "banner": {
+    //js: banner
   },
   format: "cjs",
-  external: ["vscode", "lodash"],
+  external: ["vscode"],
   loader: {
     ".ts": "ts",
     ".js": "js",
-    ".tsx": "tsx"
+    ".cjs": "js",
+    ".tsx": "tsx",
+    ".css": "css",
+    ".ttf": "file"
   },
   logLevel: "info",
-  sourcemap: "linked",
+  sourcemap: "inline",
   define: {
     // Reemplaza process.env.NODE_ENV por "production" (o "development")
     "process.env.NODE_ENV": "\"production\""
   },
+  plugins: [
+    postCssPlugin({
+      postcss: {
+        plugins: [require('@tailwindcss/postcss')],
+      }
+    }),
+  ]
 };
 
 export default config;
+