+++ tailwind.config.js
@@ -1,0 +1,9 @@
+
+
+/** @type {import('tailwindcss').Config} */
+export default {
+    content: ['./src/**/*.{js,jsx,ts,tsx}'],
+    // ...
+    // The rest of the tailwindcss configuration
+    // For more, see: https://tailwindcss.com/docs/configuration
+};